TST: tweak fixtures a bit
 - fix clean up of databases
 - give names to map db -> fixture
 - move all to conftest.py
